ciencia-da-computacao/project-inventory-report/dev-requirements.txt
-r requirements.txt

wheel==0.37.1
black==22.3.0
factory-boy==3.2.1
Faker==13.4.0
flake8==4.0.1
pytest==7.0.1
pytest-cov==2.10.1
pytest-json==0.4.0
xmltodict==0.13.0
git+https://github.com/betrybe/pytest-dependency
